+++ src/inputs/index.js
@@ -1,0 +1,3 @@
+import { cacheInput, getCachedInput, inputIsCached } from './inputCache.js';
+
+export { cacheInput, getCachedInput, inputIsCached };
